+++ parser.h
@@ -13,6 +13,6 @@ gchar* parser_valid_target(const gchar *target);
 int parser_int(const gchar *str, int typ, const gchar *option, t_minmax range);
 const char* parser_str(const gchar *str, const gchar *option, int buff_sz, int rx_ndx);
 void parser_whois(gchar *buff, int sz, gchar* elem[]);
-t_minmax parser_range(gchar *range);
+t_minmax parser_range(gchar *range, const gchar *option);
 
 #endif