+++ ui/appbar.c
@@ -6,7 +6,6 @@
 #include "pinger.h"
 #include "parser.h"
 #include "notifier.h"
-#include "tabs/graph.h"
 
 #define ENTER_HINT "Enter hostname or IP address ..."
 
@@ -20,7 +19,6 @@ static int update_datetime(gpointer label) {
     time_t now = time(NULL);
     strftime(datetime_label, sizeof(datetime_label), "%F %T", localtime(&now));
     gtk_label_set_text(GTK_LABEL(label), datetime_label);
-    if (opts.graph) graphtab_update();
   }
   return datetime_id;
 }